+++ geonode/contrib/edit_data/tests.py
@@ -1,3 +1,47 @@
+# -*- coding: utf-8 -*-
+#########################################################################
+#
+# Copyright (C) 2017 OSGeo
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+#########################################################################
+
+from geonode.tests.base import GeoNodeBaseTestSupport
+
+import dj_database_url
+
+from django.conf import settings
+from django.core.urlresolvers import reverse
 from django.test import TestCase
 
-# Create your tests here.
+from guardian.shortcuts import get_anonymous_user
+
+from geonode.geoserver.signals import gs_catalog
+from geonode import GeoNodeException
+from geonode.layers.models import Layer
+from geonode.layers.populate_layers_data import create_layer_data
+
+
+
+
+class EditDataCoreTest(GeoNodeBaseTestSupport):
+
+    def setUp(self):
+        super(EditDataCoreTest, self).setUp()
+
+        create_layer_data()
+        self.user = 'admin'
+        self.passwd = 'admin'
+        self.anonymous_user = get_anonymous_user()
